requirements.txt
openai
streamlit
python-dotenv
pymupdf
python-docx
pytesseract
pillow
tinydb
pydantic
reportlab